crud solucion errores

+++ src/controllers/hotel.controller.js
@@ -6,32 +6,26 @@ const underscore = require('underscore')
 // Agregar
 function agregarHoteles(req, res) {
     let parametros = req.body
-    if (parametros.nombre && parametros.direccion && parametros.pais && parametros.dueno) {
-        Usuario.findOne({ nombre: parametros.dueno }, (err, usuarioEncontrado) => {
-            if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
-            if (underscore.isEmpty(usuarioEncontrado)) return res.status(404).send({ mensaje: 'Datos vacios' })
-            if (usuarioEncontrado.rol !== 'ROL_ADMIN') {
-                return res.status(404).send({ mensaje: 'Este usuario no es dueno' })
-            } else {
-                Hotel.find({ nombre: parametros.nombre }, (err, hotelesEncontrados) => {
-                    if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
-                    if (underscore.isEmpty(hotelesEncontrados)) {
-                        let hotelModel = new Hotel();
-                        hotelModel.nombre = parametros.nombre
-                        hotelModel.direccion = parametros.direccion
-                        hotelModel.pais = parametros.pais
-                        hotelModel.idAdmin = usuarioEncontrado._id
-                        hotelModel.save((err, hotelGuardado) => {
-                            if (err) return res.status(404).send({ mensaje: 'Error en la peticion al guardar' });
-                            return res.status(200).send({ mensaje: hotelGuardado })
-                        })
+    if (parametros.nombre && parametros.direccion && parametros.pais) {
 
-                    } else {
-                        return res.status(200).send({ mensaje: 'Ya existe un hotel con ese nombre' })
-                    }
+        Hotel.find({ nombre: parametros.nombre }, (err, hotelesEncontrados) => {
+            if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
+            if (underscore.isEmpty(hotelesEncontrados)) {
+                let hotelModel = new Hotel();
+                hotelModel.nombre = parametros.nombre
+                hotelModel.direccion = parametros.direccion
+                hotelModel.pais = parametros.pais
+                hotelModel.idAdmin = req.user.sub
+                hotelModel.save((err, hotelGuardado) => {
+                    if (err) return res.status(404).send({ mensaje: 'Error en la peticion al guardar' });
+                    return res.status(200).send({ mensaje: hotelGuardado })
                 })
+
+            } else {
+                return res.status(200).send({ mensaje: 'Ya existe un hotel con ese nombre' })
             }
         })
+
     } else {
         return res.status(500).send({ mensaje: 'No has llenado todos los campos' })
     }
@@ -41,7 +35,7 @@ function agregarHoteles(req, res) {
 function editarHoteles(req, res) {
     let idH = req.params.idHotel;
     let parametros = req.body
-    if (parametros.dueno) {
+    if (parametros.dueno || parametros.idAdmin) {
         return res.status(500).send({ mensaje: 'No se puede cambiar ese campo' })
     } else {
         Hotel.findById(idH, (err, hotelesEncontrados) => {
@@ -49,10 +43,14 @@ function editarHoteles(req, res) {
             if (underscore.isEmpty(hotelesEncontrados)) {
                 return res.status(404).send({ mensaje: 'No existe un hotel con ese id' })
             } else {
-                Hotel.findByIdAndUpdate(idH, parametros, { new: true }, (err, hotelEditado) => {
-                    if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
-                    return res.status(200).send({ mensaje: hotelEditado })
-                })
+                if (hotelesEncontrados.idAdmin === req.user.sub) {
+                    Hotel.findByIdAndUpdate(idH, parametros, { new: true }, (err, hotelEditado) => {
+                        if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
+                        return res.status(200).send({ mensaje: hotelEditado })
+                    })
+                } else {
+                    return res.status(404).send({ mensaje: 'No posees un hotel con ese id' })
+                }
             }
         })
     }
@@ -65,10 +63,14 @@ function eliminarHoteles(req, res) {
         if (underscore.isEmpty(hotelesEncontrados)) {
             return res.status(404).send({ mensaje: 'No existe un hotel con ese id' })
         } else {
-            Hotel.findByIdAndDelete(idH, (err, hotelEditado) => {
-                if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
-                return res.status(200).send({ mensaje: hotelEditado })
-            })
+            if (hotelesEncontrados.idAdmin === req.user.sub) {
+                Hotel.findByIdAndDelete(idH, (err, hotelEditado) => {
+                    if (err) return res.status(404).send({ mensaje: 'Error en la peticion al encontrar' });
+                    return res.status(200).send({ mensaje: hotelEditado })
+                })
+            } else {
+                return res.status(404).send({ mensaje: 'No posees un hotel con ese id' })
+            }
         }
     })
 }
